cloud/workers/requirements.txt
# Python Worker Dependencies
# Stage 6: Python Worker Integration

# HTTP requests for LLM APIs
requests>=2.31.0

# YAML parsing (for config if needed)
pyyaml>=6.0

# Token counting for OpenAI models
tiktoken>=0.5.0

# Environment variable loading
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Type checking
mypy>=1.0.0
types-requests>=2.31.0
types-pyyaml>=6.0
